chore(storage): Re-factored Storage instance management to improve tree-shaking.
