+++ packages/storage/src/Storage.ts
@@ -11,7 +11,7 @@
  * and limitations under the License.
  */
 
-import { ConsoleLogger as Logger, Parser } from '@aws-amplify/core';
+import { Amplify, ConsoleLogger as Logger, Parser } from '@aws-amplify/core';
 import { AWSS3Provider } from './providers';
 import {
 	StorageCopySource,
@@ -35,6 +35,7 @@ import { PutObjectCommandInput } from '@aws-sdk/client-s3';
 import { AWSS3UploadTask } from './providers/AWSS3UploadTask';
 
 const logger = new Logger('StorageClass');
+const loggerStorageInstance = new Logger('Storage'); // Logging relating to Storage instance management
 
 const DEFAULT_PROVIDER = 'AWSS3';
 /**
@@ -404,6 +405,41 @@ export class Storage {
 		return prov.list(path, config) as StorageListOutput<T>;
 	}
 }
+
+/**
+ * Configure & register Storage singleton instance.
+ */
+let _instance: Storage = null;
+const getInstance = () => {
+	if (_instance) {
+		return _instance;
+	}
+	loggerStorageInstance.debug('Create Storage Instance, debug');
+	_instance = new Storage();
+	_instance.vault = new Storage();
+
+	const old_configure = _instance.configure;
+	_instance.configure = options => {
+		loggerStorageInstance.debug('storage configure called');
+		const vaultConfig = { ...old_configure.call(_instance, options) };
+
+		// set level private for each provider for the vault
+		Object.keys(vaultConfig).forEach(providerName => {
+			if (typeof vaultConfig[providerName] !== 'string') {
+				vaultConfig[providerName] = {
+					...vaultConfig[providerName],
+					level: 'private',
+				};
+			}
+		});
+		loggerStorageInstance.debug('storage vault configure called');
+		_instance.vault.configure(vaultConfig);
+	};
+	return _instance;
+};
+
+export const StorageInstance: Storage = getInstance();
+Amplify.register(Storage);
 
 /**
  * @deprecated use named import
