+++ packages/storage/src/index.ts
@@ -11,50 +11,13 @@
  * and limitations under the License.
  */
 
-import { Storage as StorageClass } from './Storage';
-
-import { Amplify, ConsoleLogger as Logger } from '@aws-amplify/core';
-
-const logger = new Logger('Storage');
-
-let _instance: StorageClass = null;
-
-const getInstance = () => {
-	if (_instance) {
-		return _instance;
-	}
-	logger.debug('Create Storage Instance, debug');
-	_instance = new StorageClass();
-	_instance.vault = new StorageClass();
-
-	const old_configure = _instance.configure;
-	_instance.configure = options => {
-		logger.debug('storage configure called');
-		const vaultConfig = { ...old_configure.call(_instance, options) };
-
-		// set level private for each provider for the vault
-		Object.keys(vaultConfig).forEach(providerName => {
-			if (typeof vaultConfig[providerName] !== 'string') {
-				vaultConfig[providerName] = {
-					...vaultConfig[providerName],
-					level: 'private',
-				};
-			}
-		});
-		logger.debug('storage vault configure called');
-		_instance.vault.configure(vaultConfig);
-	};
-	return _instance;
-};
-
-export const Storage: StorageClass = getInstance();
-Amplify.register(Storage);
+import { Storage, StorageInstance } from './Storage';
 
 /**
  * @deprecated use named import
  */
-export default Storage;
+export default StorageInstance;
 
-export { StorageClass };
+export { Storage as StorageClass, StorageInstance as Storage };
 export * from './providers';
 export * from './types';
